prime factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  *
  * @num: number checked
  */
-void largest_prime(int num)
+void largest_prime(long int num)
 {
 	int i, prime;
